feat(backend): use MDC to include the organism in log messages #542
