+++ backend/src/main/kotlin/org/pathoplexus/backend/config/WebConfig.kt
@@ -1,16 +1,22 @@
 package org.pathoplexus.backend.config
 
+import org.pathoplexus.backend.log.OrganismMdcInterceptor
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @Configuration
-class CorsConfiguration : WebMvcConfigurer {
+class WebConfig : WebMvcConfigurer {
     override fun addCorsMappings(registry: CorsRegistry) {
         registry.addMapping("/**")
             .allowedOrigins("*") // Allow requests from any origin
             .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
             .allowedHeaders("*")
             .maxAge(3600) // Max age of pre-flight requests
+    }
+
+    override fun addInterceptors(registry: InterceptorRegistry) {
+        registry.addInterceptor(OrganismMdcInterceptor())
     }
 }
