--- backend/src/main/kotlin/org/pathoplexus/backend/log/ResponseLogger.kt
@@ -1,23 +1,0 @@
-package org.pathoplexus.backend.log
-
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import mu.KotlinLogging
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-
-private val log = KotlinLogging.logger {}
-
-@Component
-class ResponseLogger : OncePerRequestFilter() {
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain,
-    ) {
-        filterChain.doFilter(request, response)
-
-        log.info { "${request.requestURL} - Responding with status ${response.status}" }
-    }
-}
