+++ backend/src/test/resources/logback-test.xml
@@ -1,7 +1,7 @@
 <configuration>
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
         <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>%date %level [%thread] %class: %message%n</Pattern>
+            <Pattern>%date %level [%thread] [%X{organism}] - %class: %message%n</Pattern>
         </layout>
     </appender>
